Release 1.4.3

+++ jni/emerald/items/EmeraldDoorItem.cpp
@@ -2,7 +2,7 @@
 #include "mcpe/world/material/Material.h"
 #include "mcpe/CreativeItemCategory.h"
 
-EmeraldDoorItem::EmeraldDoorItem(short itemId) : DoorItem("emeraldDoor", itemId, Material::getMaterial(MaterialType::METAL), DoorBlock::DoorType(0)){
+EmeraldDoorItem::EmeraldDoorItem(int itemId) : DoorItem("emeraldDoor", itemId, Material::getMaterial(MaterialType::METAL), DoorBlock::DoorType(0)){
 	setMaxStackSize(16);
 	creativeCategory = CreativeItemCategory::DECORATIONS;
 }
